Creation of required interfaces and classes ready for implementation of generics.
