+++ Src/BootCamp.Chapter/ISubject.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using static BootCamp.Chapter.ITeacher;
+
+namespace BootCamp.Chapter
+{
+    interface ISubject
+    {
+
+        // Leave it empty, because subjects are unrelated. Just for simulation
+        // Alternatively use base Subject class.
+        interface ISubject { }
+
+    }
+}
