+++ Src/BootCamp.Chapter/ITeacher.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter
+{
+    interface ITeacher
+    {
+        public interface ITeacher<TSubject> where TSubject : ISubject
+        {
+            TSubject ProduceMaterial();
+        }
+    }
+}
