+++ Src/BootCamp.Chapter/Schools/HighSchool.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter.Schools
+{
+    class HighSchool : ISchool
+    {
+    }
+}
