+++ Src/BootCamp.Chapter/Schools/MiddleSchool.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter.Schools
+{
+    class MiddleSchool : ISchool
+    {
+    }
+}
