+++ Src/BootCamp.Chapter/Schools/University.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter.Schools
+{
+    class University : ISchool
+    {
+    }
+}
