+++ Src/BootCamp.Chapter/Students/HighSchoolStudent.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter.Students
+{
+    class HighSchoolStudent : IStudent
+    {
+    }
+}
