+++ Src/BootCamp.Chapter/Students/UniversityStudent.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter.Students
+{
+    class UniversityStudent : IStudent
+    {
+    }
+}
