+++ Src/BootCamp.Chapter/Subjects/Maths.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter.Subjects
+{
+    class Maths : ISubject
+    {
+    }
+}
