+++ Src/BootCamp.Chapter/Subjects/Programming.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter.Subjects
+{
+    class Programming : ISubject
+    {
+    }
+}
