+++ Src/BootCamp.Chapter/Teachers/HighSchoolTeacher.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter.Teachers
+{
+    class HighSchoolTeacher : ITeacher
+    {
+    }
+}
