+++ Src/BootCamp.Chapter/Teachers/MiddleSchoolTeacher.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter.Teachers
+{
+    class MiddleSchoolTeacher : ITeacher
+    {
+    }
+}
