+++ Src/BootCamp.Chapter/Teachers/UniversityTeacher.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BootCamp.Chapter.Teachers
+{
+    class UniversityTeacher : ITeacher
+    {
+    }
+}
